0003-gitlab-rest-api/pyproject.toml
[project]
name = "0003-gitlab-rest-api"
version = "0.0.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
license = {file = "LICENSE"}
keywords = []
authors = [
  { name = "Muhammad Hafiz bin Muhamad", email = "mhafiz.muhamad@gmail.com" }
]
classifiers = [
  "Private :: Do not Upload",
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3.13",
]
dependencies = [
  "httpx",
  "python-dotenv",
]

# [project.scripts]

# [project.urls]

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.basedpyright]
reportOptionalMemberAccess = "none"
typeCheckingMode = "standard"
